+++ src/App.scss
@@ -80,7 +80,3 @@
 .error-message {
   color: var(--persimmon);
 }
-
-.no-username-status-message {
-  color: var(--white);
-}
