+++ src/pages/Insights/index.js
@@ -1,9 +1,12 @@
 import React from 'react'
 import cx from 'classnames'
-import { Route, Switch, Redirect } from 'react-router-dom'
+import { connect } from 'react-redux'
+import { Route, Switch, Redirect, Link } from 'react-router-dom'
 import Loadable from 'react-loadable'
 import { Helmet } from 'react-helmet'
 import PageLoader from '../../components/Loader/PageLoader'
+import { showNotification } from './../../actions/rootActions'
+import { getTimeIntervalFromToday } from './../../utils/dates'
 import styles from './index.module.scss'
 
 export const baseLocation = '/insights'
@@ -28,7 +31,29 @@ const LoadableUnAuthPage = Loadable({
   loading: () => <PageLoader />
 })
 
-const PageHub = ({ location: { pathname }, isLoggedIn, isDesktop }) => {
+const PageHub = ({
+  location: { pathname },
+  isLoggedIn,
+  isDesktop,
+  showUsernameAlert,
+  hasUsername
+}) => {
+  const setLastShowUsernameAlert = () => {
+    showUsernameAlert()
+    localStorage.setItem('withoutUsernameSince', new Date())
+  }
+
+  if (isLoggedIn && !hasUsername) {
+    console.log('here')
+    const lastShowDate = Date.parse(
+      localStorage.getItem('withoutUsernameSince')
+    )
+    if (lastShowDate) {
+      const { from } = getTimeIntervalFromToday(-7, 'd')
+      if (lastShowDate - from < 0) setLastShowUsernameAlert()
+    } else setLastShowUsernameAlert()
+  }
+
   window.scrollTo(0, 0)
   const normalizedPathname = pathname.endsWith('/')
     ? pathname.slice(0, -1)
@@ -92,4 +117,36 @@ const PageHub = ({ location: { pathname }, isLoggedIn, isDesktop }) => {
   )
 }
 
-export default PageHub
+const mapStateToProps = ({ user: { data } }) => ({
+  userId: data.id,
+  hasUsername: !!data.username
+})
+
+const mapDispatchToProps = dispatch => ({
+  showUsernameAlert: () => {
+    dispatch(
+      showNotification({
+        variant: 'warning',
+        solidFill: true,
+        title: 'Please add a username',
+        description:
+          'We need to know more about you to release additional functionality',
+        actions: [
+          {
+            label: (
+              <Link className={styles.usernameLink} to='/account'>
+                Add username
+              </Link>
+            )
+          }
+        ],
+        dismissAfter: 80000
+      })
+    )
+  }
+})
+
+export default connect(
+  mapStateToProps,
+  mapDispatchToProps
+)(PageHub)
